requirements.txt
aiohttp==1.0.5
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cloud-init==18.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
discord.py==0.16.12
distro-info==0.18
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
language-selector==0.1
MarkupSafe==1.0
multidict==4.3.1
oauthlib==2.0.6
PAM==0.4.2
pbr==4.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pygobject==3.26.1
PyJWT==1.5.3
PyNaCl==1.0.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.2
python-debian==0.1.32
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
service-identity==16.0.0
six==1.11.0
ssh-import-id==5.7
stevedore==1.29.0
systemd-python==234
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
websockets==3.4
zope.interface==4.3.2
